+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package com.leeyiyuan;
+package com.leeyiyuan.task;
 
 public class Task {
